requirements.txt
asgiref==3.2.4
Django==3.0.4
numpy==1.18.1
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
torch==1.4.0
torchvision==0.5.0
